refactor: use Robin Wieruch's approach to have kebab-case for file names instead of ugly PascalCase
